third_party/WebKit/LayoutTests/fast/compositor-wheel-scroll-latching/non-animated-scroll/subpixel-accumulation.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
  let t;
  if (window.internals && chrome && chrome.gpuBenchmarking) {
    internals.settings.setPreferCompositingToLCDTextEnabled(true);
    internals.settings.setScrollAnimatorEnabled(false);
    t = async_test("Vertical scroll shouldn't latch to horizontal scroller.");
  }

  // Helper async function to block execution for n number of rAFs.
  async function nFrames(n) {
    return new Promise(resolve => {
      let remainingFrames = n;
      let func = function() {
        --remainingFrames;
        if (remainingFrames === 0)
          resolve();
        else {
          requestAnimationFrame(func);
        }
      };

      if (n === 0) {
        resolve();
      } else {
        requestAnimationFrame(() => {
          func(resolve);
        });
      }
    });
  }

  async function scroll(pixels_to_scroll, direction, start_x, start_y, speed_in_pixels_s) {
    return new Promise((resolve, reject) => {
      const WHEEL_SOURCE_TYPE = 2; // MOUSE_INPUT from synthetic_gesture_params.h
      chrome.gpuBenchmarking.smoothScrollBy(pixels_to_scroll,
                                            resolve,
                                            start_x,
                                            start_y,
                                            WHEEL_SOURCE_TYPE,
                                            direction,
                                            speed_in_pixels_s);
    });
  }

  async function runTest() {
    if (!chrome.gpuBenchmarking)
      return;

    // Wait a bit to ensure we don't start sending scroll gestures to the
    // compositor before the page layers have fully propagated to the CC active
    // tree.
    await nFrames(2);

    // Scroll horizontally first. Ensure we scrolled the div since otherwise we
    // may pass the test vacuously. See crbug.com/801381 for the case we're
    // guarding against.
    await scroll(10, 'right', 50, 50, 1000);

    // Wait a bit since the scroll gesture resolves when the ScrollEnd is sent
    // from the browser, not when it's processed by the renderer.
    await nFrames(2);

    t.step(() => {
      // Approx because smoothScrollBy isn't perfectly accurate.
      assert_approx_equals(
          document.getElementById("scroller").scrollLeft, 10, 1,
          "Sanity check, if this fails the test isn't in the correct state.");
    });

    // Now scroll vertically. The scroller shouldn't have any vertical scroll
    // extent so the window should scroll.
    await scroll(10, 'down', 50, 50, 1000);
    await nFrames(2);
    t.step(() => {
      assert_approx_equals(window.scrollY, 10, 1);
    });
    t.done();
  }

  addEventListener('load', runTest);
</script>
<style>
  ::-webkit-scrollbar {
    width: 0px;
    height: 0px;
  }
  body {
    margin: 0;
    height: 2000px;
  }
  #scroller {
    position: relative;
    top: 0.5625px;
    width: 200px;
    height: 200.8125px;
    overflow: auto;
  }
  #space {
    width: 1000px;
    background-color: coral;
    height: 200.8125px;
  }
</style>
<div id="scroller">
  <div id="space"></div>
</div>
